Draft to not disclose information to the end user when requesting a password reset, a reconfirmation or an unlocking.
This is not a final code, the `Application.get_env(:haytni, :mode)` will be replaced
in the future with some *config* key and this *strict mode* is not (yet) tested (and
disabled by default).

Also, this "feature" is debatable: is it really relevant? If you use email addresses,
and for example, don't tell when requesting a password reset if it exists or not but
your registering process discloses it...